feat: Payment 구현 및 할인기능 구현
Payment 현금/카드 등 지불수단에 대한 Enum 클래스 구현
Menu isChicken 메서드 추가
Customer Payment를 받아 이에 따른 할인 적용
Table 만약 치킨 메뉴가 10개 이상일 시 할인 적용

+++ src/main/java/domain/money/Money.java
@@ -22,8 +22,8 @@ public class Money {
         return money;
     }
 
-    public Money multiply(int operand) {
-        return new Money(money * operand);
+    public Money multiply(double operand) {
+        return new Money((int)(money * operand));
     }
 
     public Money subtract(int operand) {
